requirements.txt
biopython
ncbi-datasets-cli
pandas
scikit-learn